Move double buffer into render chooser
